July 25,  2023 1:37 AM
I removed the Agenda column in the Activities table. I also removed the start_datetime and end_datetime in the requirement_bins table. I also added the deadline column to the bin table.

I added annimation in deleting a file in the upload_modal of the faculty

I fixed some bugs in activities module in the staff role

I also added the name of the creator of the activitiy or requirement bin in the bin_details and activity_details module.